git-svn-id: svn://hamptoncatlin.com/haml/tags/stable@713 7063305b-7217-0410-af8c-cdc13e5119b9

+++ lib/sass/plugin.rb
@@ -26,62 +26,40 @@ module Sass
       def options=(value)
         @@options.merge!(value)
       end
-      
+
       # Checks each stylesheet in <tt>options[:css_location]</tt>
       # to see if it needs updating,
       # and updates it using the corresponding template
       # from <tt>options[:templates]</tt>
       # if it does.
       def update_stylesheets
+        return if options[:never_update]
+
         Dir.glob(File.join(options[:template_location], "**", "*.sass")).entries.each do |file|
-          
+
           # Get the relative path to the file with no extension
           name = file.sub(options[:template_location] + "/", "")[0...-5]
-                    
-          if options[:always_update] || stylesheet_needs_update?(name)
+
+          if !forbid_update?(name) && (options[:always_update] || stylesheet_needs_update?(name))
             css = css_filename(name)
             File.delete(css) if File.exists?(css)
-            
+
             filename = template_filename(name)
             l_options = @@options.dup
             l_options[:filename] = filename
-            l_options[:load_paths] = (l_options[:load_paths] || []) + [l_options[:template_location]]
+            l_options[:load_paths] = load_paths
             engine = Engine.new(File.read(filename), l_options)
-            begin
-              result = engine.render
-            rescue Exception => e
-              if options[:full_exception]
-                e_string = "#{e.class}: #{e.message}"
+            result = begin
+                       engine.render
+                     rescue Exception => e
+                       exception_string(e)
+                     end
 
-                if e.is_a? Sass::SyntaxError
-                  e_string << "\non line #{e.sass_line}"
-
-                  if e.sass_filename
-                    e_string << " of #{e.sass_filename}"
-
-                    if File.exists?(e.sass_filename)
-                      e_string << "\n\n"
-
-                      min = [e.sass_line - 5, 0].max
-                      File.read(e.sass_filename).rstrip.split("\n")[
-                          min .. e.sass_line + 5
-                      ].each_with_index do |line, i|
-                        e_string << "#{min + i + 1}: #{line}\n"
-                      end
-                    end
-                  end
-                end
-                result = "/*\n#{e_string}\n\nBacktrace:\n#{e.backtrace.join("\n")}\n*/"
-              else
-                result = "/* Internal stylesheet error */"
-              end
-            end
-            
             # Create any directories that might be necessary
             dirs = [l_options[:css_location]]
             name.split("/")[0...-1].each { |dir| dirs << "#{dirs[-1]}/#{dir}" }
             dirs.each { |dir| Dir.mkdir(dir) unless File.exist?(dir) }
-            
+
             # Finally, write the file
             File.open(css, 'w') do |file|
               file.print(result)
@@ -89,19 +67,76 @@ module Sass
           end
         end
       end
-      
+
       private
-      
+
+      def load_paths
+        (options[:load_paths] || []) + [options[:template_location]]
+      end
+
+      def exception_string(e)
+        if options[:full_exception]
+          e_string = "#{e.class}: #{e.message}"
+
+          if e.is_a? Sass::SyntaxError
+            e_string << "\non line #{e.sass_line}"
+
+            if e.sass_filename
+              e_string << " of #{e.sass_filename}"
+
+              if File.exists?(e.sass_filename)
+                e_string << "\n\n"
+
+                min = [e.sass_line - 5, 0].max
+                File.read(e.sass_filename).rstrip.split("\n")[
+                  min .. e.sass_line + 5
+                ].each_with_index do |line, i|
+                  e_string << "#{min + i + 1}: #{line}\n"
+                end
+              end
+            end
+          end
+          "/*\n#{e_string}\n\nBacktrace:\n#{e.backtrace.join("\n")}\n*/"
+        else
+          "/* Internal stylesheet error */"
+        end
+      end
+
       def template_filename(name)
-        "#{@@options[:template_location]}/#{name}.sass"
+        "#{options[:template_location]}/#{name}.sass"
       end
-      
+
       def css_filename(name)
-        "#{@@options[:css_location]}/#{name}.css"
+        "#{options[:css_location]}/#{name}.css"
       end
-      
+
+      def forbid_update?(name)
+        name[0] == ?_
+      end
+
       def stylesheet_needs_update?(name)
-        !File.exists?(css_filename(name)) || (File.mtime(template_filename(name)) - 2) > File.mtime(css_filename(name))
+        if !File.exists?(css_filename(name))
+          return true
+        else
+          css_mtime = File.mtime(css_filename(name))
+          File.mtime(template_filename(name)) > css_mtime ||
+            dependencies(template_filename(name)).any?(&dependency_updated?(css_mtime))
+        end
+      end
+
+      def dependency_updated?(css_mtime)
+        lambda do |dep|
+          File.mtime(dep) > css_mtime ||
+            dependencies(dep).any?(&dependency_updated?(css_mtime))
+        end
+      end
+
+      def dependencies(filename)
+        File.readlines(filename).grep(/^@import /).map do |line|
+          line[8..-1].split(',').map do |inc|
+            Sass::Engine.find_file_to_import(inc.strip, load_paths)
+          end
+        end.flatten.grep(/\.sass$/)
       end
     end
   end
